legacy/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
python-multipart>=0.0.6

# MCP SDK
git+https://github.com/modelcontextprotocol/python-sdk.git

# Image processing
pillow>=9.5.0
opencv-python>=4.7.0

# HTTP client
requests>=2.28.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
jinja2>=3.1.2
numpy>=1.24.0
uuid>=1.30.0
tqdm>=4.65.0

# Image Generation - Venice.ai API (optional)
# (using existing requests and python-dotenv)

# Image Generation - Google Gemini API
google-generativeai>=0.3.0

# Network and Service Discovery
zeroconf>=0.39.0
aiohttp>=3.8.4

# 3D Reconstruction - CUDA Multi-View Stereo
open3d>=0.17.0
trimesh>=3.21.0
pyrender>=0.1.45

# Remote Processing
fastapi-utils>=0.2.1
python-jose>=3.3.0  # For JWT authentication
aiofiles>=23.1.0

# For development
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
pytest-asyncio>=0.21.0

# Deprecated dependencies (kept for reference)
# segment-anything-2>=1.0
# torch>=2.0.0
# torchvision>=0.15.0
# pytorch3d>=0.7.4
# ninja>=1.11.0
